requirements.txt
Flask
gunicorn
ngram
gevent
esy-osm-pbf
